Moved tests into their own files
