+++ tests/ErrorTest.php
@@ -1,0 +1,58 @@
+<?php
+class ErrorTest extends PHPUnit_Framework_TestCase {
+
+    public function testNoTSSRules() {
+		$this->expectException("Exception");
+		$this->expectExceptionMessage("No TSS rules parsed");
+		$template = new \Transphporm\Builder("<div></div>", "NONEXISTANT_FILE");
+		$template->output();
+	}
+
+	public function testFunctionException() {
+		$this->expectException("Transphporm\\Exception");
+		$this->expectExceptionMessage("TSS error: Problem carrying out function 'data' on Line 1 of tss");
+
+		$xml = '<div></div>';
+		$tss = '
+		div { content: data(getTest()); }
+		';
+		$template = new \Transphporm\Builder($xml, $tss);
+		$template->output();
+	}
+
+    public function testPseudoException() {
+		$this->expectException("Transphporm\\Exception");
+		$this->expectExceptionMessage("TSS error: Problem carrying out pseudo 'nth-child' on Line 1 of tss");
+
+		$xml = '<div></div>';
+		$tss = '
+		div:nth-child(h) { content: "test"; }
+		';
+		$template = new \Transphporm\Builder($xml, $tss);
+		$template->output();
+	}
+
+    public function testFormatterException() {
+		$this->expectException("Transphporm\\Exception");
+		$this->expectExceptionMessage("TSS error: Problem carrying out formatter 'test' on Line 1 of tss");
+
+		$xml = '<div></div>';
+		$tss = '
+		div { content: "test"; format: test; }
+		';
+		$template = new \Transphporm\Builder($xml, $tss);
+		$template->output();
+	}
+
+    public function testPropertyException() {
+		$this->expectException("Transphporm\\Exception");
+		$this->expectExceptionMessage("TSS error: Problem carrying out property 'repeat' on Line 1 of tss");
+
+		$xml = '<div></div>';
+		$tss = '
+		div { repeat: "test"; }
+		';
+		$template = new \Transphporm\Builder($xml, $tss);
+		$template->output();
+	}
+}
