+++ tests/TransphpormTest.php
@@ -402,309 +402,6 @@ class TransphpormTest extends PHPUnit_Framework_TestCase {
 		</div>'), $this->stripTabs($template->output()->body));
 	}
 
-	public function testBefore() {
-		$template =  '
-		<div>Test</div>
-		';
-
-		$tss = 'div:before {content: "BEFORE";}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<div>BEFORETest</div>'), $this->stripTabs($template->output()->body));
-	}
-
-	public function testAfter() {
-		$template =  '
-		<div>Test</div>
-		';
-
-		$tss = 'div:after {content: "AFTER";}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<div>TestAFTER</div>'), $this->stripTabs($template->output()->body));
-	}
-
-
-	public function testIterationPseudo() {
-		$data = new stdclass;
-		$data->list = [];
-
-		$one = new stdclass;
-		$one->name = 'One';
-		$one->id = '1';
-		$data->list[] = $one;
-
-		$two = new stdclass;
-		$two->name = 'Two';
-		$two->id = '2';
-		$data->list[] = $two;
-
-		$three = new stdclass;
-		$three->name = 'Three';
-		$three->id = '3';
-		$data->list[] = $three;
-
-
-		$template = '
-				<ul>
-					<li>
-						<h2>header</h2>
-						<span>TEST1</span>
-					</li>
-				</ul>
-		';
-
-		$css = 'ul li {repeat: data(list);}
-		ul li h2 {content: iteration(id)}
-		ul li span {content: iteration(name); }
-		ul li span:iteration[id="2"] {display: none;}
-		';
-
-
-		$template = new \Transphporm\Builder($template, $css);
-
-
-		$this->assertEquals($this->stripTabs('<ul>
-			<li>
-				<h2>1</h2>
-				<span>One</span>
-			</li><li>
-				<h2>2</h2>
-
-			</li><li>
-				<h2>3</h2>
-				<span>Three</span>
-			</li>
-		</ul>'), $this->stripTabs($template->output($data)->body));
-
-	}
-
-
-	public function testIterationPseudoNotEquals() {
-		$data = new stdclass;
-		$data->list = [];
-
-		$one = new stdclass;
-		$one->name = 'One';
-		$one->id = '1';
-		$data->list[] = $one;
-
-		$two = new stdclass;
-		$two->name = 'Two';
-		$two->id = '2';
-		$data->list[] = $two;
-
-		$three = new stdclass;
-		$three->name = 'Three';
-		$three->id = '3';
-		$data->list[] = $three;
-
-
-		$template = '
-				<ul>
-					<li>
-						<h2>header</h2>
-						<span>TEST1</span>
-					</li>
-				</ul>
-		';
-
-		$css = 'ul li {repeat: data(list);}
-		ul li h2 {content: iteration(id)}
-		ul li span {content: iteration(name); }
-		ul li span:iteration[id!="2"] {display: none;}
-		';
-
-
-		$template = new \Transphporm\Builder($template, $css);
-
-
-		$this->assertEquals($this->stripTabs('<ul>
-			<li>
-				<h2>1</h2>
-			</li><li>
-				<h2>2</h2>
-				<span>Two</span>
-			</li><li>
-				<h2>3</h2>
-			</li>
-		</ul>'), $this->stripTabs($template->output($data)->body));
-
-	}
-	public function testMultiPseudo() {
-		$data = new stdclass;
-		$data->list = [];
-
-		$one = new stdclass;
-		$one->name = 'One';
-		$one->id = '1';
-		$data->list[] = $one;
-
-		$two = new stdclass;
-		$two->name = 'Two';
-		$two->id = '2';
-		$data->list[] = $two;
-
-		$three = new stdclass;
-		$three->name = 'Three';
-		$three->id = '3';
-		$data->list[] = $three;
-
-
-		$template = '
-				<ul>
-					<li>
-						<h2>header</h2>
-						<span>TEST1</span>
-					</li>
-				</ul>
-		';
-
-		$css = 'ul li {repeat: data(list);}
-		ul li h2 {content: iteration(id)}
-		ul li span {content: iteration(name); }
-		ul li span:iteration[id="2"]:before {content: "BEFORE";}
-		';
-
-
-		$template = new \Transphporm\Builder($template, $css);
-
-
-		$this->assertEquals($this->stripTabs('<ul>
-			<li>
-				<h2>1</h2>
-				<span>One</span>
-			</li><li>
-				<h2>2</h2>
-				<span>BEFORETwo</span>
-			</li><li>
-				<h2>3</h2>
-				<span>Three</span>
-			</li>
-		</ul>'), $this->stripTabs($template->output($data)->body));
-
-	}
-
-	public function testNthChild() {
-		$template = '
-			<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>
-		';
-
-		$tss = 'ul li:nth-child(2) {content: "REPLACED"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<ul>
-				<li>One</li>
-				<li>REPLACED</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>'), $this->stripTabs($template->output()->body));
-
-
-	}
-
-	public function testNthChild2() {
-		$template = '
-			<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>
-		';
-
-		$tss = 'ul li:nth-child(3) {content: "REPLACED"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>REPLACED</li>
-				<li>Four</li>
-			</ul>'), $this->stripTabs($template->output()->body));
-
-	}
-
-	public function testNthChild3() {
-		$template = '
-			<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>
-		';
-
-		$tss = 'ul li:nth-child(1) {content: "REPLACED"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<ul>
-				<li>REPLACED</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>'), $this->stripTabs($template->output()->body));
-
-	}
-
-
-	public function testNthChildOdd() {
-		$template = '
-
-			<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>
-		';
-
-		$tss = 'ul li:nth-child(odd) {content: "REPLACED"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<ul>
-				<li>REPLACED</li>
-				<li>Two</li>
-				<li>REPLACED</li>
-				<li>Four</li>
-			</ul>'), $this->stripTabs($template->output()->body));
-
-	}
-
-	public function testNthChildEven() {
-		$template = '
-			<ul>
-				<li>One</li>
-				<li>Two</li>
-				<li>Three</li>
-				<li>Four</li>
-			</ul>
-		';
-
-		$tss = 'ul li:nth-child(even) {content: "REPLACED"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<ul>
-				<li>One</li>
-				<li>REPLACED</li>
-				<li>Three</li>
-				<li>REPLACED</li>
-			</ul>'), $this->stripTabs($template->output()->body));
-
-	}
-
 	public function testReadAttribute() {
 		$template = '
 			<div class="fromattribute">Test</div>
@@ -781,123 +478,6 @@ div {content: "bar"; }
 		$this->assertEquals('<div>bar</div>', $template->output()->body);
 	}
 
-	public function testImport() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$file = __DIR__ . DIRECTORY_SEPARATOR . 'import.tss';
-		$file = str_replace('\\', '/', $file);
-		$tss = "
-			@import '$file';
-		";
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals('<div>foo</div>', $template->output()->body);
-	}
-
-	public function testImportDynamic() {
-		$template = '
-			<div>Test</div>
-		';
-
-
-		$tss = "
-			@import data(filename);
-		";
-
-		$data = [
-			'filename' => __DIR__  . DIRECTORY_SEPARATOR . 'import.tss'
-		];
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals('<div>foo</div>', $template->output($data)->body);
-	}
-
-	public function testImportMiddleOfFile() {
-		$template = '
-			<span>foo</span>
-			<div>Test</div>
-			<h1>foo</h1>
-		';
-
-		$file = __DIR__ . DIRECTORY_SEPARATOR . 'import.tss';
-		$file = str_replace('\\', '/', $file);
-		$tss = "
-			span {content: 'test1';}
-			@import '$file';
-			h1 {content: 'h1';}
-		";
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<span>test1</span><div>foo</div><h1>h1</h1>'), $this->stripTabs($template->output()->body));
-	}
-
-	public function testMultiImport() {
-		$template = '
-			<span>test</span>
-			<div>Test</div>
-			<h1>foo</h1>
-		';
-
-		$file = __DIR__ . DIRECTORY_SEPARATOR . 'import.tss';
-		$file2 = __DIR__ . DIRECTORY_SEPARATOR . 'import2.tss';
-		$file = str_replace('\\', '/', $file);
-		$file2 = str_replace('\\', '/', $file2);
-		$tss = "
-			span {content: 'test1';}
-			@import '$file';
-			h1 {content: 'h1';}
-			@import '$file2';
-		";
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<span>bar</span><div>foo</div><h1>h1</h1>'), $this->stripTabs($template->output()->body));
-
-	}
-
-
-	public function testMultiImportTogether() {
-		$template = '
-			<span>test</span>
-			<div>Test</div>
-			<h1>foo</h1>
-		';
-
-		$file = __DIR__ . DIRECTORY_SEPARATOR . 'import.tss';
-		$file2 = __DIR__ . DIRECTORY_SEPARATOR . 'import2.tss';
-		$file = str_replace('\\', '/', $file);
-		$file2 = str_replace('\\', '/', $file2);
-		$tss = "
-			span {content: 'test1';}
-			@import '$file';
-			@import '$file2';
-			h1 {content: 'h1';}
-
-		";
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<span>bar</span><div>foo</div><h1>h1</h1>'), $this->stripTabs($template->output()->body));
-
-	}
-
-	public function testImportInImportedFile() {
-		$template = new \Transphporm\Builder("tests/test.xml", "tests/other/otherImport.tss");
-
-		$this->assertEquals('<!DOCTYPE html><html><body>test</body></html>', $this->stripTabs($template->output()->body));
-	}
-
-	public function testBaseDirChangeWithImport() {
-		$template = new \Transphporm\Builder("tests/test.xml", "tests/other/templateFromImport.tss");
-
-		$this->assertEquals('<!DOCTYPE html><html><body><p>foo</p></body></html>', $this->stripTabs($template->output()->body));
-	}
-
 	public function testContentTemplate() {
 		$template = '
 			<div>Test</div>
@@ -955,7 +535,6 @@ div {content: "bar"; }
 		</form>'), $this->stripTabs($template->output($data)->body));
 	}
 
-
 	public function testBind() {
 		//Binds a specific pice of data (which may be an array to an element)
 		//All calls to data() for child elements will use this as the root data object
@@ -983,108 +562,6 @@ div {content: "bar"; }
 		</form>'), $this->stripTabs($template->output($data)->body));
 
 	}
-
-	public function testFormatNumber() {
-		$template = '
-			<div>Test</div>
-		';;
-		$tss = 'div {content: "1.234567"; format: decimal 2;}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals('<div>1.23</div>', $this->stripTabs($template->output()->body));
-
-
-	}
-
-	public function testFormatCurrency() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$tss = 'div {content: "1.234567"; format: currency;}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals('<div>£1.23</div>', $this->stripTabs($template->output()->body));
-
-
-	}
-
-
-	public function testFormatCurrencyAfter() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$tss = 'div {content: "1.234567"; format: currency;}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-		$locale = json_decode(file_get_contents('src/Formatter/Locale/enGB.json'), true);
-		$locale['currency_position'] = 'after';
-		$template->loadModule(new \Transphporm\Module\Format($locale));
-
-		$this->assertEquals('<div>1.23£</div>', $this->stripTabs($template->output()->body));
-	}
-
-	public function testFormatStringUpper() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$tss = 'div {content: "test"; format: uppercase}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-		$template->loadModule(new \Transphporm\Module\Format);
-		$this->assertEquals('<div>TEST</div>', $this->stripTabs($template->output()->body));
-
-
-	}
-
-
-	public function testFormatStringLower() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$tss = 'div {content: "test"; format: lowercase}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-		$template->loadModule(new \Transphporm\Module\Format);
-		$this->assertEquals('<div>test</div>', $this->stripTabs($template->output()->body));
-
-
-	}
-
-	public function testCustomFormatter() {
-		$template = '
-			<div>Test</div>
-		';
-
-		$tss = 'div {content: "test"; format: reverse}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-		require_once 'tests/ReverseFormatter.php';
-		$template->loadModule(new ReverseFormatterModule);
-
-		$this->assertEquals('<div>tset</div>', $this->stripTabs($template->output()->body));
-
-	}
-
-	public function testFormatStringTitle() {
-		$template = '
-			<div>test</div>
-		';
-
-		$tss = 'div {content: "a test title"; format: titlecase}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals('<div>A Test Title</div>', $this->stripTabs($template->output()->body));
-
-
-	}
-
 
 	public function testSemicolonInString() {
 
@@ -1197,47 +674,6 @@ div {content: "bar"; }
 
 		$this->assertEquals($this->stripTabs('<foo xmlns="http://foo/"><bar>yy</bar></foo>'), $this->stripTabs($template->output()->body));
 	}
-
-	public function testFormatDate() {
-
-		$template = '
-			<div>test</div>
-		';
-
-		$tss = 'div {content: "2015-10-05"; format: date}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-
-		$this->assertEquals($this->stripTabs('<div>05/10/2015</div>'), $this->stripTabs($template->output()->body));
-	}
-
-	public function testFormatDateCustom() {
-
-		$template = '
-			<div>test</div>
-		';
-
-		$tss = 'div {content: "2015-10-05"; format: date "jS M"}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-
-		$this->assertEquals($this->stripTabs('<div>5th Oct</div>'), $this->stripTabs($template->output()->body));
-	}
-
-	public function testHTMLFormat() {
-		$template = '
-			<div>test</div>
-		';
-
-		$tss = 'div {content: "<span>foobar</span>"; format: html}';
-
-		$template = new \Transphporm\Builder($template, $tss);
-
-		$this->assertEquals($this->stripTabs('<div><span>foobar</span></div>'), $this->stripTabs($template->output()->body));
-	}
-
 
 	public function testIncludeTemplatePartial() {
 		$template = '
@@ -1475,41 +911,6 @@ select option[value=data()]:attr(selected) { content: "selected"; }
 		</select>'));
 	}
 
-	public function testOverrideAfter() {
-			$xml = '<div>
-			<span>Test</span>
-		</div>';
-
-
-		$includeFile = __DIR__ . DIRECTORY_SEPARATOR . 'include.xml';
-		$includeFile = str_replace('\\', '/', $includeFile);
-
-		$tss = "div:after {content: 'foo' }
-div:after {content: 'bar' }
-		";
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals('<div><span>Test</span>bar</div>', $this->stripTabs($template->output()->body));
-
-	}
-
-
-	public function testOverrideBefore() {
-			$xml = '<div>
-			<span>Test</span>
-		</div>';
-
-
-		$includeFile = __DIR__ . DIRECTORY_SEPARATOR . 'include.xml';
-
-		$tss = "div:before {content: 'foo' }
-		div:before {content: 'bar';}
-		";
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals('<div>bar<span>Test</span></div>', $this->stripTabs($template->output()->body));
-
-	}
 	public function testAttrDisplayNone() {
 		$xml = '<div>
 			<span class="bar">test</span>
@@ -1569,70 +970,6 @@ div:after {content: 'bar' }
 		'));
 	}
 
-	public function testFilterDataArray() {
-		$data = [
-			'anArray' => [
-				'one' => 'foo',
-				'two' => 'bar'
-			]
-		];
-
-		$xml = '
-		<div class="one">
-
-		</div>
-		<div class="two">
-
-		</div>
-		<div class="three">
-		</div>
-		';
-
-		$tss = 'div:data[anArray[attr("class")]] {content: "set"; }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs('
-			<div class="one">set</div>
-			<div class="two">set</div>
-			<div class="three"></div>
-		'), $this->stripTabs($template->output($data)->body));
-
-	}
-
-	public function testFilterDataArrayWithEquals() {
-		$data = [
-			'anArray' => [
-				'one' => 'foo',
-				'two' => 'bar'
-			]
-		];
-
-		$xml = '
-		<div class="one">
-
-		</div>
-		<div class="two">
-
-		</div>
-		<div class="three">
-		</div>
-		';
-
-		$tss = 'div:data[anArray[attr(class)]="foo"] {content: data(anArray[attr(class)]) }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output($data)->body), $this->stripTabs('
-			<div class="one">foo</div>
-			<div class="two"></div>
-			<div class="three">
-			</div>
-		'));
-
-	}
-
-
 	public function testMultiRule() {
 		$xml = '
 		<div class="one">
@@ -1655,57 +992,6 @@ div:after {content: 'bar' }
 			<div class="three">foo</div>
 		'));
 
-
-	}
-
-	public function testNot() {
-		$xml = '
-		<div class="one">
-
-		</div>
-		<div class="two">
-
-		</div>
-		<div class="three">
-		</div>
-		';
-
-		$tss = 'div:not(.two) {content: "foo"; }';
-
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output()->body), $this->stripTabs('
-			<div class="one">foo</div>
-			<div class="two"></div>
-			<div class="three">foo</div>
-		'));
-
-	}
-
-	public function testNotWithMoreDepth() {
-		$xml = '
-		<div class="one">
-			<p></p>
-		</div>
-		<div class="two">
-			<p></p>
-		</div>
-		<div class="three">
-			<p></p>
-		</div>
-		';
-
-		$tss = 'p:not("div.two p") {content: "foo"; }';
-
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output()->body), $this->stripTabs('
-			<div class="one"><p>foo</p></div>
-			<div class="two"><p></p></div>
-			<div class="three"><p>foo</p></div>
-		'));
 
 	}
 
@@ -1835,69 +1121,6 @@ ul li span {
 		$this->assertEquals($this->stripTabs($template->output($obj)->body), $this->stripTabs('<div><p>foo</p></div>'));
 
 	}
-
-	public function testAttrPseudoSelectorAlternateSyntax() {
-		$xml = '<div></div>';
-
-		$tss = 'div:[data()=true] {content: "test" }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output(true)->body), $this->stripTabs('<div>test</div>'));
-	}
-
-	public function testFunctionCallAsConditonal() {
-
-		$xml = '<div></div>';
-
-		$obj = new Foo();
-
-		$tss = 'div:data[returnTrue()=true] {content: "test" }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output($obj)->body), $this->stripTabs('<div>test</div>'));
-	}
-
-	public function testFunctionCallAsConditonal2() {
-
-		$xml = '<div></div>';
-
-		$obj = new Foo();
-
-		$tss = 'div:data[returnFalse()=false] {content: "test" }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output($obj)->body), $this->stripTabs('<div>test</div>'));
-	}
-
-	public function testFunctionCallAsConditonal3() {
-
-	    $xml = '<div></div>';
-
-	    $obj = new Foo();
-
-	    $tss = 'div:data[returnOne()=1] {content: "test" }';
-
-	    $template = new \Transphporm\Builder($xml, $tss);
-
-	    $this->assertEquals($this->stripTabs($template->output($obj)->body), $this->stripTabs('<div>test</div>'));
-	}
-
-	public function testFunctionCallAsConditonal4() {
-
-	    $xml = '<div></div>';
-
-	    $obj = new Foo();
-
-	    $tss = 'div:[data(returnOne())=1] {content: "test" }';
-
-	    $template = new \Transphporm\Builder($xml, $tss);
-
-	    $this->assertEquals($this->stripTabs($template->output($obj)->body), $this->stripTabs('<div>test</div>'));
-	}
-
 
 	public function testRuleOneComment() {
 		$xml = '<div></div>';
@@ -2048,36 +1271,6 @@ ul li span {
 		$this->assertEquals($this->stripTabs('<div>bar</div>'), $this->stripTabs($template->output($data)->body));
 	}
 
-	public function testNl2brBasic() {
-        $xml = "
-        <div></div>
-        ";
-
-        $tss = "
-        div { content: 'Test Line 1 \n Test Line 2'; format: nl2br; }
-        ";
-
-        $transphporm = new Builder($xml, $tss);
-
-        $this->assertEquals($this->stripTabs('<div>Test Line 1 <br /> Test Line 2</div>'), $this->stripTabs($transphporm->output()->body));
-    }
-
-    public function testNl2brBasicFromData() {
-        $xml = "
-        <div></div>
-        ";
-
-        $tss = "
-        div { content: data(); format: nl2br; }
-        ";
-
-        $data = "Test Line 1 \n Test Line 2";
-
-        $transphporm = new Builder($xml, $tss);
-
-        $this->assertEquals($this->stripTabs('<div>Test Line 1 <br /> Test Line 2</div>'), $this->stripTabs($transphporm->output($data)->body));
-    }
-
 	public function testRoot() {
 		$xml = "
 		<div></div>
@@ -2095,19 +1288,6 @@ ul li span {
 		$template = new \Transphporm\Builder($xml, $tss);
 
 		$this->assertEquals($this->stripTabs('<div>bar1bar2</div>'), $this->stripTabs($template->output($data)->body));
-	}
-
-	public function testEmptyFunctionCallAsConditonal() {
-
-		$xml = '<div></div>';
-
-		$data = "test";
-
-		$tss = 'div:[data()="test"] {content: "test" }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($this->stripTabs($template->output($data)->body), $this->stripTabs('<div>test</div>'));
 	}
 
 	public function testRepeatNotExistant() {
@@ -2138,41 +1318,6 @@ ul li span {
 		$template = new \Transphporm\Builder($xml, $tss);
 
 		$this->assertEquals('<div>Test2</div>', $template->output()->body);
-	}
-
-	public function testNoTSSRules() {
-		$this->expectException("Exception");
-		$this->expectExceptionMessage("No TSS rules parsed");
-		$template = new \Transphporm\Builder("<div></div>", "NONEXISTANT_FILE");
-		$template->output();
-	}
-
-	public function testNthChild4() {
-		$xml = '
-		<h2 class="name">Test</h2>
-		<h3 class="name">Test</h3>
-		<h4 class="name">Test</h4>
-		';
-
-		$tss = '.name:nth-child(2) { display: none; }
-
-		.name:nth-child(3) { display: none; }';
-
-		$template = new \Transphporm\Builder($xml, $tss);
-
-		$this->assertEquals($template->output()->body, $this->stripTabs('<h2 class="name">Test</h2>'));
-	}
-
-	public function testTransphpormException() {
-		$this->expectException("Transphporm\\Exception");
-		$this->expectExceptionMessage("TSS error: Problem carrying out function 'data' on Line 1 of tss");
-
-		$xml = '<div></div>';
-		$tss = '
-		div { content: data(getTest()); }
-		';
-		$template = new \Transphporm\Builder($xml, $tss);
-		$template->output();
 	}
 }
